+++ module-activity-service/src/main/java/com/joosangah/activityservice/common/client/FeignClientConfiguration.java
@@ -1,4 +1,4 @@
-package com.joosangah.newsfeedservice;
+package com.joosangah.activityservice.common.client;
 
 import feign.RequestInterceptor;
 import feign.RequestTemplate;
