+++ module-activity-service/src/main/java/com/joosangah/activityservice/common/client/NewsfeedFeignClient.java
@@ -1,4 +1,4 @@
-package com.joosangah.activityservice;
+package com.joosangah.activityservice.common.client;
 
 import com.joosangah.activityservice.post.domain.entity.Post;
 import org.springframework.cloud.openfeign.FeignClient;
