+++ module-newsfeed-service/src/main/java/com/joosangah/newsfeedservice/common/client/FeignClientConfiguration.java
@@ -1,4 +1,4 @@
-package com.joosangah.userservice;
+package com.joosangah.newsfeedservice.common.client;
 
 import feign.RequestInterceptor;
 import feign.RequestTemplate;
