+++ module-newsfeed-service/src/main/java/com/joosangah/newsfeedservice/common/client/UserFeignClient.java
@@ -1,5 +1,6 @@
-package com.joosangah.newsfeedservice;
+package com.joosangah.newsfeedservice.common.client;
 
+import com.joosangah.newsfeedservice.common.client.FeignClientConfiguration;
 import com.joosangah.newsfeedservice.common.domain.User;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
