+++ module-user-service/src/main/java/com/joosangah/userservice/common/client/FeignClientConfiguration.java
@@ -1,4 +1,4 @@
-package com.joosangah.activityservice;
+package com.joosangah.userservice.common.client;
 
 import feign.RequestInterceptor;
 import feign.RequestTemplate;
