+++ module-user-service/src/main/java/com/joosangah/userservice/mail/MailService.java
@@ -1,4 +1,4 @@
-package com.joosangah.userservice;
+package com.joosangah.userservice.mail;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
